01_stt_resumo_llm/requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
python-multipart
jinja2
